update_script_vnext checkpoint
